Fixed Floor
Added small 'sidewalk' walls that stop the player.

This is to prevent the player from falling off the platform horizontally.

Also stretched the street vertically very long, to prevent players from falling off the platform to the left or right (unless they are bored and choose to walk that distance).